requirements.txt
# requirements.txt — Pain Miner dependencies

# AI
google-genai>=1.0.0

# Reddit scraping
praw>=7.7.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Scheduling (for future cron mode)
apscheduler>=3.10.1
